### 2025-05-17 #8240 @editorialbot
[PRE REVIEW]: Den2Obj: A command line tool for producing isosurfaces from electron density data files
**Submitting author:** @ifilot (<a href="http://orcid.org/0000-0003-1403-8379">Ivo Filot</a>)
**Repository:** https://github.com/ifilot/den2obj
**Branch with paper.md** (empty if default branch): paper
**Version:** v1.2.2
