### 2025-09-06 #8240 @HaoZeke
hi @imtambovtcev @utf @arosen93 👋 would you be interested in and available to review this JOSS submission? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
