### 2026-02-17 #8240 @HaoZeke
@NicoRenaud just following up on the reviewer invitation from a few weeks ago — are you available to review this submission?

@ifilot in the meantime, it would be great if you could start addressing the preliminary feedback from @imtambovtcev regarding the scope/novelty framing and the installation issues on Ubuntu/macOS.
