### 2026-02-17 #8240 @HaoZeke
hi @muralidhar-nalabothula @raghurama123 @AlexBuccheri 👋 would you be interested in and available to review this JOSS submission? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
