### 2026-02-17 #8240 @HaoZeke
@AlexBuccheri I'm happy to add a third if that works, the more the merrier :)
