### 2026-02-21 #8240 @AlexBuccheri
> [@AlexBuccheri](https://github.com/AlexBuccheri) I'm happy to add a third if that works, the more the merrier :)

Ok, I'm happy to look at the paper, build, documentation and example cases.
