### 2026-03-05 #10100 @imtambovtcev
For now, I haven't seen any changes in the project related to my prereview, nor have I received any comments from the author, so I will repost my prereview here as the current state review of the submission.

As stated in the article, there are multiple codes that can create isosurfaces from electronic structure data, including mine. Functioning from the terminal or being able to save to common 3D object formats is a nice and important feature, but it is also present in some other codes. Another feature of this code is the new file format d2o, which is stated to be memory efficient. This is not surprising, since a compressed binary format is naturally smaller than human-readable formats, although it comes at the cost of readability. Yet I feel really optimistic about the increase in the variety of visualization codes, as it is what gives us flexibility in the workflow, and I highly value this. At the same time, since some of the functionality overlaps with existing tools, it would be helpful if the author could more clearly highlight what is truly novel or distinctive about Den2Obj, and whether this is the kind of contribution that fits well within the scope of JOSS.

I had some problems with the installation: it failed on Ubuntu and macOS, but I was able to use the code with Debian. The tests in the workflows prove that Ubuntu installation is possible, so the problem is probably with my device/setup. Still, it would be better to have the Docker file updated so that it can handle the setup properly, because right now it does not. That would help with the integration.
