### 2026-03-06 0305eda
Revising compilation procedure
