+++ .github/workflows/build-openvdb.yml
@@ -8,20 +8,31 @@ on:
 
 jobs:
   build:
-
     runs-on: ubuntu-latest
 
     steps:
-    - uses: actions/checkout@v3
-    - name: Install dependencies
-      run: sudo apt install -y build-essential cmake libglm-dev libtclap-dev libboost-all-dev libopenvdb-dev libtbb-dev libcppunit-dev libeigen3-dev wget  liblzma-dev zlib1g-dev libbz2-dev
-    - name: Get sources OpenVDB 8.2
-      run: wget https://github.com/AcademySoftwareFoundation/openvdb/archive/refs/tags/v8.2.0.tar.gz && tar -xvzf v8.2.0.tar.gz
-    - name: Compile and install OpenVDB 8.02
-      run: mkdir openvdb-build && cd openvdb-build && cmake ../openvdb-8.2.0 -DCMAKE_INSTALL_PREFIX=/opt/openvdb && make -j3 && sudo make install
-    - name: Configure CMake
-      run: mkdir build && cd build && cmake -DMOD_OPENVDB=1 ../src
-    - name: Build
-      run: cd build && make -j3
-    - name: Perform unit tests
-      run: cd build && make test
+      - uses: actions/checkout@v4
+      - name: Install dependencies
+        run: |
+          sudo apt update
+          sudo apt install -y \
+            build-essential \
+            cmake \
+            libglm-dev \
+            libtclap-dev \
+            libboost-all-dev \
+            libopenvdb-dev \
+            libtbb-dev \
+            libcppunit-dev \
+            libeigen3-dev \
+            liblzma-dev \
+            zlib1g-dev \
+            libbz2-dev \
+            libssl-dev \
+            pkg-config
+      - name: Configure CMake
+        run: cmake -S src -B build -DMOD_OPENVDB=ON
+      - name: Build
+        run: cmake --build build --parallel
+      - name: Run unit tests
+        run: ctest --test-dir build --output-on-failure
