+++ src/CMakeLists.txt
@@ -18,138 +18,124 @@
  #                                                                        *
  #*************************************************************************/
 
-# set minimum cmake requirements
-cmake_minimum_required(VERSION 3.5)
-project (den2obj)
+cmake_minimum_required(VERSION 3.16)
+project(den2obj VERSION 1.2.2 LANGUAGES CXX)
 
-# change compiler directives when code coverage is required
-if(USE_GCOV)
-  set(CMAKE_BUILD_TYPE "Debug")
+option(USE_GCOV "Build with gcov coverage instrumentation" OFF)
+option(MOD_OPENVDB "Enable OpenVDB support" OFF)
+option(HAS_OPENMP "Enable OpenMP support when available" ON)
+option(DISABLE_TEST "Disable building tests" OFF)
 
-  # Set global c and c++ flags
-  set(CMAKE_C_FLAGS "${CMAKE_C_FLAGS} -fprofile-arcs -ftest-coverage")
-  set(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -fprofile-arcs -ftest-coverage")
-
-  # Link flags used for creating executables
-  set(CMAKE_EXE_LINKER_FLAGS "${CMAKE_EXE_LINKER_FLAGS} -lgcov -fprofile-arcs -ftest-coverage")
-
-  # Link flags used for creating shared libraries
-  set(CMAKE_SHARED_LINKER_FLAGS "${CMAKE_SHARED_LINKER_FLAGS} -lgcov -ftest-coverage")
+if(NOT CMAKE_BUILD_TYPE AND NOT CMAKE_CONFIGURATION_TYPES)
+    message(STATUS "Setting build type to 'Release' as none was specified.")
+    set(CMAKE_BUILD_TYPE Release CACHE STRING "Choose the type of build." FORCE)
+    set_property(CACHE CMAKE_BUILD_TYPE PROPERTY STRINGS "Debug" "Release" "MinSizeRel" "RelWithDebInfo")
 endif()
 
-# add custom directory to look for .cmake files
-set(CMAKE_MODULE_PATH ${CMAKE_MODULE_PATH} ${CMAKE_SOURCE_DIR}/cmake/modules )
+set(PROGNAME "DEN2OBJ")
+set(VERSION_MAJOR "${PROJECT_VERSION_MAJOR}")
+set(VERSION_MINOR "${PROJECT_VERSION_MINOR}")
+set(VERSION_MICRO "${PROJECT_VERSION_PATCH}")
 
-# store git hash
 execute_process(
     COMMAND git log -1 --format=%h
     WORKING_DIRECTORY ${CMAKE_CURRENT_LIST_DIR}
     OUTPUT_VARIABLE GIT_HASH
     OUTPUT_STRIP_TRAILING_WHITESPACE
+    ERROR_QUIET
 )
 
-# prepare configuration file
-SET(PROGNAME "DEN2OBJ")
-SET(VERSION_MAJOR "1")
-SET(VERSION_MINOR "2")
-SET(VERSION_MICRO "2")
-message(STATUS "Writing configuration file in: ${CMAKE_CURRENT_SOURCE_DIR}/config.h")
-configure_file(${CMAKE_CURRENT_SOURCE_DIR}/config.h.in ${CMAKE_CURRENT_SOURCE_DIR}/config.h @ONLY)
-configure_file(${CMAKE_CURRENT_SOURCE_DIR}/den2obj.pc.in ${CMAKE_BINARY_DIR}/den2obj.pc @ONLY)
[…]

[…]
