### 2026-03-07 2bd1ca6
Replace fragile MD5 file checks with structural output validation
