+++ src/test/test_file_creation.cpp
@@ -29,106 +29,134 @@ void TestFileCreation::tearDown() {
 }
 
 void TestFileCreation::test_ply_file() {
-    // set number of threads to 1
-    omp_set_num_threads(1);
-
-    // read scalar field
-    ScalarField sf("co_2pi_x.cub", ScalarFieldInputFileType::SFF_CUB);
-    sf.read();
-    
-    // perform marching cubes algorithm
-    IsoSurface is(&sf);
-    is.marching_cubes(0.01);
-    
-    // construct mesh
-    IsoSurfaceMesh ism(&sf, &is);
-    ism.construct_mesh(true);
-
-    // create file
-    ism.write_to_file("test.ply", "co molecule 2pi_x orbital test", "molecule");
-
-    // test md5sum
-    CPPUNIT_ASSERT_EQUAL(this->md5("test.ply"), std::string("ec143f6ebf9b72761803c3b091f54bf3"));
+    const auto ref = this->generate_mesh();
+    this->assert_ply_shape("test.ply", ref);
 }
 
 void TestFileCreation::test_stl_file() {
-    // set number of threads to 1
-    omp_set_num_threads(1);
-
-    // read scalar field
-    ScalarField sf("co_2pi_x.cub", ScalarFieldInputFileType::SFF_CUB);
-    sf.read();
-    
-    // perform marching cubes algorithm
-    IsoSurface is(&sf);
-    is.marching_cubes(0.01);
-    
-    // construct mesh
-    IsoSurfaceMesh ism(&sf, &is);
-    ism.construct_mesh(true);
-
-    // create file
-    ism.write_to_file("test.stl", "co molecule 2pi_x orbital test", "molecule");
-
-    // test md5sum
-    CPPUNIT_ASSERT_EQUAL(this->md5("test.stl"), std::string("c2194ba639caf5092654862bb9f93298"));
+    const auto ref = this->generate_mesh();
+    this->assert_stl_shape("test.stl", ref);
 }
 
 void TestFileCreation::test_obj_file() {
+    const auto ref = this->generate_mesh();
+    this->assert_obj_shape("test.obj", ref);
+}
+
+TestFileCreation::MeshReference TestFileCreation::generate_mesh() const {
     // set number of threads to 1
     omp_set_num_threads(1);
 
     // read scalar field
     ScalarField sf("co_2pi_x.cub", ScalarFieldInputFileType::SFF_CUB);
     sf.read();
-    
+
     // perform marching cubes algorithm
     IsoSurface is(&sf);
     is.marching_cubes(0.01);
-    
+
     // construct mesh
     IsoSurfaceMesh ism(&sf, &is);
     ism.construct_mesh(true);
 
-    // create file
+    // create files for all supported output formats
+    ism.write_to_file("test.ply", "co molecule 2pi_x orbital test", "molecule");
+    ism.write_to_file("test.stl", "co molecule 2pi_x orbital test", "molecule");
     ism.write_to_file("test.obj", "co molecule 2pi_x orbital test", "molecule");
 
-    // test md5sum
-    CPPUNIT_ASSERT_EQUAL(this->md5("test.obj"), std::string("e2b3e09f9c010dac99a7bc0137c187ec"));
+    return {
+        ism.get_vertices().size(),
+        ism.get_normals().size(),
+        ism.get_texcoords().size() / 6
+    };
 }
 
-/**
- * @brief      Calculate MD5 checksum of a file
- *
- * @param[in]  name      Path to the file
- * 
- * @return     32 byte string containing md5 checksum
- */
-std::string TestFileCreation::md5(const std::string& filename) {
-    // read the file
[…]
