### 2026-03-08 #10100 @ifilot
@imtambovtcev Thank you very much for taking the time to review the submission and for reposting your prereview comments here. I appreciate the constructive feedback. I had a few questions with respect to the comments you raised and hope you are able to point me in the right direction such that I can effectively implement your feedback.

---

## 1. Comparison with existing tools

You mentioned that there are multiple codes capable of generating isosurfaces from electronic structure data, including one you developed. I would be very interested in learning more about your implementation so that I can better position Den2Obj relative to existing tools. If the code is publicly available, could you please share a reference (for example a repository or publication)? I went through your repos, but could not find anything related to isosurface generation.

---

## 2. Clarifying the unique contribution of Den2Obj

In the manuscript, the following aspects are already mentioned:

- A command-line oriented workflow designed for automated pipelines rather than interactive GUI environments.
- Direct support for commonly used electronic-structure output formats such as VASP `CHGCAR`, `PARCHG`, and Gaussian `Cube`.
- Conversion to common mesh formats (`.obj`, `.ply`, `.stl`) for downstream visualization workflows.
- A compressed native scalar-field format (`.d2o`) intended for efficient storage of electron-density grids.
- Parallel implementations of the marching cubes and marching tetrahedra algorithms using OpenMP.
- Support for dual isosurface extraction (positive and negative lobes) for wavefunction visualization.
- Optional export to OpenVDB for volumetric rendering workflows.

Based on your comment, this is apparently insufficient. I can of course extend the list further, yet are there any specific aspects you are looking for? Perhaps this also becomes apparent when I compare my implementation to yours (point 1).

---

## 3. Installation issues

Thank you also for testing the installation on multiple systems. Since the GitHub Actions workflows demonstrate successful builds on Ubuntu, it would be very helpful to better understand what went wrong in your environment. Would it be possible to either share the error messages you encountered, or  open a GitHub issue describing the installation failure (including OS version, compiler version, and build logs)? This would allow me to reproduce the problem and improve the installation process and Docker configuration accordingly.

---

Thank you again for the thoughtful review and for taking the time to test the software on multiple platforms.
