### 2026-03-30 #10100 @imtambovtcev
Dear @ifilot,

Thank you for your detailed response, and I sincerely apologize for the long delay, the past weeks have been quite intense on my end.

1. My code
The repository is ChemParse/ChemVista. Its main focus is on the visualization of molecular geometry dynamics, but it can handle isosurface generation and conversion to .stl and multiple other formats (via PyVista) as well.

2. Installation issues
Since my pre-review, I have had a substantial change in my system environment. I attempted the installation again on Ubuntu, and the build completed successfully; all tests passed. I can no longer reproduce the problem I originally encountered, so my Ubuntu claim is no longer valid. I apologize for the confusion.

3. Unique contribution
I will leave the final judgment on the scope and novelty of the contribution to the other reviewers and the editor. From my perspective, the .d2o compressed binary format is the genuinely new element introduced by this work. The other features, like command-line isosurface extraction, export to .obj/.ply/.stl, support for VASP and Gaussian formats, are certainly useful and well-implemented, but they are implementations of established ideas available in existing tools in one way or another. I don't say this to diminish the work - having another tool in this space is valuable, but in terms of what is new to the field, I mostly see the .d2o format itself.
