### 2026-05-21 8fb5042
Strengthening README.md file
