+++ README.md
@@ -8,121 +8,220 @@
 [![License: GPL v3](https://img.shields.io/badge/License-GPLv3-blue.svg)](https://www.gnu.org/licenses/gpl-3.0)
 
 ## Purpose
-Den2Obj is a command-line tool that construct isosurfaces from densely packed
-scalar fields. Den2Obj supports VASP charge files such as CHGCAR and PARCHG,
-Gaussian .cube files as well as its own .d2o format.
+
+Den2Obj is a command-line tool for computational chemists and materials
+scientists who want to visualize volumetric scalar fields from ab initio
+calculations as 3D isosurface meshes.
+
+It reads scalar field data produced by [VASP](https://www.vasp.at/) (CHGCAR,
+PARCHG, LOCPOT) and [Gaussian](https://gaussian.com/) (`.cub`) and extracts
+isosurfaces using the marching cubes or marching tetrahedra algorithm. The
+resulting surfaces — representing quantities such as electron density, partial
+charge density, or electrostatic potential — are written to standard 3D mesh
+formats (`.obj`, `.ply`) that can be directly imported into tools like Blender
+or MeshLab for high-quality rendering.
+
+Den2Obj also supports format conversion: VASP and Gaussian files can be
+converted to the native `.d2o` compressed binary format or to OpenVDB (`.vdb`)
+for volumetric rendering workflows. When writing `.d2o`, Den2Obj benchmarks all
+supported compression algorithms (gzip, lzma, bzip2) and automatically selects
+the one that produces the smallest file, so no manual tuning is needed.
 
 ## Example images
+
 ![Canonical valence orbitals of CH4](docs/_static/img/ch4_valence_orbitals.png)
+
+## Quick start
+
+No VASP or Gaussian files needed. Den2Obj ships with built-in datasets so you
+can try it immediately after compiling:
+
+```bash
+# Generate a built-in scalar field and extract an isosurface from it
+./den2obj -g genus2 -o genus2.d2o
+./den2obj -i genus2.d2o -o genus2.obj -v 0.5
+```
+
+The first command writes the `genus2` dataset to a `.d2o` file. The second
+extracts an isosurface at isovalue 0.5 and writes a Wavefront `.obj` mesh that
+can be opened directly in Blender or MeshLab. See
+[Built-in datasets](#built-in-dataset-generation) for the full list.
 
 ## Compilation instructions
 
-### Debian Latest
+### Debian-based systems (Ubuntu, Debian)
 
-Getting the dependencies
-
-```bash
-sudo apt install build-essential cmake libtclap-dev libboost-all-dev libopenvdb-dev libtbb-dev \
-pkg-config libcppunit-dev libeigen3-dev liblzma-dev zlib1g-dev libbz2-dev libssl-dev
-```
-
-To compile, run the following commands:
+Tested on Ubuntu 24.04 LTS (Noble) and Debian 12 (Bookworm). Install the
+dependencies:
 
 ```bash
-git clone https://github.com/ifilot/den2obj.git
-cd den2obj
-mkdir build
-cd build
-cmake -DMOD_OPENVDB=1 ../src
-make -j5
+sudo apt install build-essential cmake libtclap-dev libboost-all-dev \
+libopenvdb-dev libtbb-dev pkg-config libcppunit-dev libeigen3-dev liblzma-dev \
+zlib1g-dev libbz2-dev libssl-dev
 ```
 
-### Ubuntu Latest
-
-The stable OpenVDB library (`libopenvdb`) under Ubuntu is incompatible with the 
-Threading Building Blocks (`libtbb`) library. To solve this, manually compile 
[…]
