### 2026-05-21 eb583a5
Adding additional compression methods and polishing I/O
