+++ src/d2o_format.cpp
@@ -19,6 +19,55 @@
  **************************************************************************/
 
 #include "d2o_format.h"
+#include "cli_format.h"
+
+#include <blosc.h>
+#include <iomanip>
+#include <zstd.h>
+
+namespace {
+    std::vector<std::pair<D2OFormat::CompressionAlgo, std::string>> ordered_compression_algos() {
+        return {
+            {D2OFormat::CompressionAlgo::GZIP, "gzip"},
+            {D2OFormat::CompressionAlgo::LZMA, "lzma"},
+            {D2OFormat::CompressionAlgo::BZIP2, "bzip2"},
+            {D2OFormat::CompressionAlgo::ZSTD, "zstd"},
+            {D2OFormat::CompressionAlgo::BLOSC, "blosc"}
+        };
+    }
+
+    void print_compression_table(
+        const std::unordered_map<D2OFormat::CompressionAlgo, std::string, std::hash<D2OFormat::CompressionAlgo>>& compressed_strings,
+        D2OFormat::CompressionAlgo selected_algo,
+        size_t original_size) {
+        CLIFormat::print_section("Compression benchmark");
+        std::cout << "+--------+-----------+--------+----------+" << std::endl;
+        std::cout << "| Method | Size      | Ratio  | Status   |" << std::endl;
+        std::cout << "+--------+-----------+--------+----------+" << std::endl;
+
+        for(const auto& algo : ordered_compression_algos()) {
+            const auto got = compressed_strings.find(algo.first);
+            if(got == compressed_strings.end()) {
+                continue;
+            }
+
+            const double ratio = static_cast<double>(got->second.size()) /
+                                 static_cast<double>(original_size) * 100.0;
+
+            const std::string status = algo.first == selected_algo ? "best" : "";
+            const std::string padded_status = status + std::string(8 - status.size(), ' ');
+
+            std::cout << "| " << std::left << std::setw(6) << algo.second
+                      << " | " << std::right << std::setw(9) << CLIFormat::format_kb(got->second.size())
+                      << " | " << std::setw(5) << (boost::format("%0.2f") % ratio).str() << "%"
+                      << " | " << CLIFormat::colorize(padded_status, CLIFormat::Color::GREEN)
+                      << " |" << std::endl;
+        }
+
+        std::cout << "+--------+-----------+--------+----------+" << std::endl;
+        CLIFormat::print_kv("Selected", D2OFormat::compression_algo_name(selected_algo));
+    }
+}
 
 /**
  * @brief      Write a D2O file
@@ -39,7 +88,8 @@ void D2OFormat::write_d2o_file(const std::string& filename,
     char buf[] = "D2O";
     outfile.write(buf, 3);
 
-    std::cout << "Writing .D2O file, using compression algo id: " << (int)algo_id << std::endl;
+    CLIFormat::print_section("Writing D2O file");
+    CLIFormat::print_kv("Compression", D2OFormat::compression_algo_name(algo_id));
 
     // copy data to char array
     size_t gridptrsz = gridptr.size() * sizeof(fpt);
@@ -49,12 +99,9 @@ void D2OFormat::write_d2o_file(const std::string& filename,
 
     std::string compressedstr;
     if(algo_id == D2OFormat::CompressionAlgo::AUTO) { // check which compression algo works best
-        // auto-look for best compression algo
-        std::cout << "Looking for best compression algorithm." << std::endl;
-
         // determine best compression format
         auto compstr = d2o_compress_all(originstr);
-        std::unordered_map<D2OFormat::CompressionAlgo, unsigned int, std::hash<D2OFormat::CompressionAlgo>> stringsizes;
[…]
