+++ docker/Dockerfile
@@ -31,6 +31,8 @@ RUN apt-get update && apt-get install -y \
   liblzma-dev \
   zlib1g-dev \
   libbz2-dev \
+  libzstd-dev \
+  libblosc-dev \
   libssl-dev
 
 RUN apt-get install curl wget

[…]
