+++ .github/workflows/build-openvdb.yml
@@ -1,38 +1,0 @@
-name: OpenVDB module
-
-on:
-  push:
-    branches: [ "master", "develop" ]
-  pull_request:
-    branches: [ "master", "develop" ]
-
-jobs:
-  build:
-    runs-on: ubuntu-latest
-
-    steps:
-      - uses: actions/checkout@v4
-      - name: Install dependencies
-        run: |
-          sudo apt update
-          sudo apt install -y \
-            build-essential \
-            cmake \
-            libglm-dev \
-            libtclap-dev \
-            libboost-all-dev \
-            libopenvdb-dev \
-            libtbb-dev \
-            libcppunit-dev \
-            libeigen3-dev \
-            liblzma-dev \
-            zlib1g-dev \
-            libbz2-dev \
-            libssl-dev \
-            pkg-config
-      - name: Configure CMake
-        run: cmake -S src -B build -DMOD_OPENVDB=ON
-      - name: Build
-        run: cmake --build build --parallel
-      - name: Run unit tests
-        run: ctest --test-dir build --output-on-failure

[…]
