### 2026-05-21 d1846b3
Patch v1.3.0 (#13)
* Revising compilation procedure

* Removing MD5 checkusum based verification

* Strengthening README.md file

* Adding additional compression methods and polishing I/O

* Improving coverage

* Updating docs
