+++ src/cli_format.cpp
@@ -1,0 +1,266 @@
+/**************************************************************************
+ *                                                                        *
+ *   Author: Ivo Filot <i.a.w.filot@tue.nl>                               *
+ *                                                                        *
+ *   DEN2OBJ is free software:                                            *
+ *   you can redistribute it and/or modify it under the terms of the      *
+ *   GNU General Public License as published by the Free Software         *
+ *   Foundation, either version 3 of the License, or (at your option)     *
+ *   any later version.                                                   *
+ *                                                                        *
+ *   DEN2OBJ is distributed in the hope that it will be useful,           *
+ *   but WITHOUT ANY WARRANTY; without even the implied warranty          *
+ *   of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.              *
+ *   See the GNU General Public License for more details.                 *
+ *                                                                        *
+ *   You should have received a copy of the GNU General Public License    *
+ *   along with this program.  If not, see http://www.gnu.org/licenses/.  *
+ *                                                                        *
+ **************************************************************************/
+
+#include "cli_format.h"
+
+#include <algorithm>
+#include <cstdlib>
+#include <iomanip>
+#include <sstream>
+#include <unistd.h>
+
+namespace {
+    /**
+     * @brief      Pad a string on the right with spaces.
+     *
+     * @param[in]  value  Value to pad
+     * @param[in]  width  Minimum output width
+     *
+     * @return     Padded string
+     */
+    std::string pad_right(const std::string& value, size_t width) {
+        if(value.size() >= width) {
+            return value;
+        }
+
+        return value + std::string(width - value.size(), ' ');
+    }
+
+    /**
+     * @brief      Get the ANSI escape sequence for a color or style.
+     *
+     * @param[in]  color  Color or style
+     *
+     * @return     ANSI escape sequence
+     */
+    std::string ansi_code(CLIFormat::Color color) {
+        switch(color) {
+            case CLIFormat::Color::BOLD:
+                return "\033[1m";
+            case CLIFormat::Color::DIM:
+                return "\033[2m";
+            case CLIFormat::Color::CYAN:
+                return "\033[36m";
+            case CLIFormat::Color::GREEN:
+                return "\033[32m";
+            case CLIFormat::Color::YELLOW:
+                return "\033[33m";
+            case CLIFormat::Color::DEFAULT:
+            default:
+                return "\033[0m";
+        }
+    }
+}
+
+/**
+ * @brief      Format a byte count as kibibytes.
+ *
+ * @param[in]  bytes  Number of bytes
+ *
+ * @return     Formatted size string
+ */
+std::string CLIFormat::format_kb(size_t bytes) {
+    std::ostringstream oss;
+    oss << std::fixed << std::setprecision(1) << (static_cast<double>(bytes) / 1024.0) << " kb";
+    return oss.str();
+}
+
+/**
+ * @brief      Format an elapsed time in seconds.
+ *
+ * @param[in]  seconds  Elapsed time in seconds
+ *
+ * @return     Formatted elapsed time string
+ */
+std::string CLIFormat::format_seconds(double seconds) {
+    std::ostringstream oss;
+    oss << std::fixed << std::setprecision(2) << seconds << " s";
+    return oss.str();
+}
+
+/**
+ * @brief      Check whether standard output is attached to a terminal.
+ *
[…]
