+++ paper/paper.bib
@@ -1,26 +1,3 @@
-@article{hafner:2008,
-    author = {Hafner, Jürgen},
-    title = {Ab-initio simulations of materials using VASP: Density-functional theory and beyond},
-    journal = {Journal of Computational Chemistry},
-    volume = {29},
-    number = {13},
-    pages = {2044-2078},
-    doi = {10.1002/jcc.21057},
-    year = {2008}
-}
-
-@article{momma:2011,
-    author = "Momma, Koichi and Izumi, Fujio",
-    title = "{{\it VESTA3} for three-dimensional visualization of crystal, volumetric and morphology data}",
-    journal = "{Journal of Applied Crystallography}",
-    year = "2011",
-    volume = "44",
-    number = "6",
-    pages = "1272-1276",
-    month = "Dec",
-    doi = {10.1107/S0021889811038970},
-}
-
 @MISC{eigenweb,
     author = {Ga\"{e}l Guennebaud and Beno\^{i}t Jacob and others},
     title = {Eigen v3},
@@ -28,25 +5,60 @@
     year = {2010}
 }
 
-@misc{bzip2web,
-    author = {Mark Wielaard and Federico Mena and Micah Snyder},
-    title = {bzip2},
-    howpublished = {https://sourceware.org/bzip2/},
-    year = {1996}
+@article{huffman:1952,
+    author = {Huffman, David A.},
+    title = {A Method for the Construction of Minimum-Redundancy Codes},
+    journal = {Proceedings of the IRE},
+    volume = {40},
+    number = {9},
+    pages = {1098--1101},
+    year = {1952},
+    doi = {10.1109/JRPROC.1952.273898}
 }
 
-@misc{gzipweb,
-    author = {Jean-loup Gailly and Mark Adler},
-    title = {gzip},
-    howpublished = {https://www.gnu.org/software/gzip/},
-    year = {1992}
+@article{ziv:1977,
+    author = {Ziv, Jacob and Lempel, Abraham},
+    title = {A Universal Algorithm for Sequential Data Compression},
+    journal = {IEEE Transactions on Information Theory},
+    volume = {23},
+    number = {3},
+    pages = {337--343},
+    year = {1977},
+    doi = {10.1109/TIT.1977.1055714}
 }
 
-@misc{lzmaweb,
-    author = {Igor Pavlov},
-    title = {lzma},
-    howpublished = {https://www.7-zip.org/},
-    year = {1996}
+@article{fenwick:1996,
+    author = {Fenwick, Peter M.},
+    title = {The Burrows-Wheeler Transform for Block Sorting Text Compression: Principles and Improvements},
+    journal = {The Computer Journal},
+    volume = {39},
+    number = {9},
+    pages = {731--740},
+    year = {1996},
+    doi = {10.1093/comjnl/39.9.731}
+}
+
+@article{alted:2010,
+    author = {Alted, Francesc},
+    title = {Why Modern CPUs Are Starving and What Can Be Done about It},
+    journal = {Computing in Science \& Engineering},
+    volume = {12},
+    number = {2},
+    pages = {68--71},
+    year = {2010},
+    doi = {10.1109/MCSE.2010.51}
+}
+
+@techreport{zstandard,
+    author = {Collet, Yann and Kucherawy, Murray S.},
+    title = {Zstandard Compression and the `application/zstd' Media Type},
+    institution = {RFC Editor},
+    type = {RFC},
+    number = {8878},
+    year = {2021},
+    month = {February},
+    doi = {10.17487/RFC8878},
+    url = {https://www.rfc-editor.org/info/rfc8878}
[…]
