+++ paper/paper.md
@@ -1,5 +1,5 @@
 ---
-title: 'Den2Obj: A command line tool for producing isosurfaces from electron density data files'
+title: 'Den2Obj: A command-line tool for producing isosurfaces from electron density data files'
 tags:
   -
 authors:
@@ -16,54 +16,42 @@ bibliography: paper.bib
 
 # Summary
 
-Electron density plays a foundational role in electronic structure calculations.
-It offers a detailed spatial representation of the distribution of electrons in
-molecules and materials. In computational methodologies, most notably Density
-Functional Theory (DFT), electron density serves as a central quantity from
-which key properties such as total energy, electrostatic potential, and atomic
-forces are derived. The analysis of electron density provides significant
-insights into chemical bonding, reactivity, and electron redistribution,
-especially during bond formation and dissociation. This capability enables the
-prediction and interpretation of the properties of materials and molecular
-systems.
+Electron density is a foundational quantity in electronic structure
+calculations. It describes the spatial distribution of electrons in molecules
+and materials, and is used to derive key properties such as total energy,
+electrostatic potential, and atomic forces. Because electron density is a
+three-dimensional scalar field, specialized visualization techniques, including
+contour plotting, isosurface generation, and volumetric rendering, are needed to
+interpret bonding, reactivity, and electron redistribution.
 
-Electron density is mathematically represented as a scalar field. It assigns a
-scalar value, corresponding to the probability density of finding an electron,
-to each point in three-dimensional space. Due to the complexity and high
-dimensionality of such data, specialized visualization techniques are necessary
-for meaningful interpretation. Common methods include contour plotting,
-isosurface generation, and volumetric rendering. These techniques facilitate the
-exploration of electronic behavior and structural characteristics.
-
-This utility, `Den2Obj`, is a C++ based command line program developed to
-generate isosurfaces from electron density scalar fields. It supports input
-formats that are commonly used in electronic structure calculations, including
-CHGCAR and PARCHG files produced by VASP, as well as Gaussian Cube files.
-`Den2Obj` is capable of converting these formats into a native `.d2o` file
-format, which achieves substantial reductions in file size through efficient
-compression. The resulting isosurfaces can be exported in several widely
+`Den2Obj` is a C++ command-line program for generating isosurfaces from electron
+density scalar fields. It supports input formats commonly used in electronic
+structure calculations, including CHGCAR and PARCHG files produced by VASP, as
+well as Gaussian Cube files. The resulting isosurfaces can be exported in widely
 supported geometry file formats, including STL (Stereolithography), PLY (Polygon
-File Format), and OBJ (Wavefront). These output formats enable convenient
-integration with a broad range of visualization and post-processing software
-tools.
+File Format), and OBJ (Wavefront), enabling integration with visualization,
+rendering, and post-processing tools. In addition, `Den2Obj` can convert scalar
+fields to a documented native `.d2o` format, a lossless binary container that
+stores the grid and unit-cell metadata while selecting an efficient compression
+method automatically.
 
 # Statement of need
-
-## Isosurface Visualization with `Den2Obj`
 
 Isosurfaces play a central role in scientific research and engineering
 applications by providing an effective means of visualizing complex scalar
 fields and understanding intricate physical phenomena. An isosurface represents
 a set of points in a field where a scalar quantity, known as the *isovalue*,
[…]
