### 2026-08-18 ifilot/den2obj#14 @jmmshn
build: error and documentation gap for OSX
Hi @ifilot,

I ran the following on my mac:

```sh
brew install cmake tclap boost openvdb tbb pkgconf cppunit eigen xz zstd c-blosc
```

and the tried the compile commands:

```sh
git clone https://github.com/ifilot/den2obj.git
cd den2obj
mkdir build && cd build
cmake -DMOD_OPENVDB=1 ../src
make -j5
```

and the `cmake` step has not worked and resulted in the following error:

```
In file included from /opt/homebrew/Cellar/eigen/5.0.1/include/eigen3/Eigen/Core:51:
/Library/Developer/CommandLineTools/SDKs/MacOSX.sdk/usr/include/c++/v1/complex:942:10: error: no member named 'hypot' in namespace 'std'; did you mean '__math::hypot'?
  942 |   return std::hypot(__c.real(), __c.imag());
      |          ^    ~~~~~
/Library/Developer/CommandLineTools/SDKs/MacOSX.sdk/usr/include/c++/v1/__math/hypot.h:34:36: note: '__math::hypot' declared here
   34 | inline _LIBCPP_HIDE_FROM_ABI float hypot(float __x, float __y) _NOEXCEPT { return __builtin_hypotf(__x, __y); }
      |                                    ^
```

It looks like this is due to the `math.h` in the code here shadowing the std `math.h`

System summary

```
macOS, Apple Silicon (arm64)
AppleClang 21.0.0.21000101  (libc++)
Homebrew: eigen 5.0.1, boost 1.92.0, cmake 4.4.2, cppunit 1.15.1,
          openvdb 13.0.0, tbb 2023.1.0, tclap 1.2.5, c-blosc 1.21.6
```
