### 2026-08-18 #10100 @jmmshn
Opened issue
https://github.com/ifilot/den2obj/issues/14
for Mac/OSX install
