### 2026-08-18 3661e9e
Hardening for MacOS
