### 2026-08-18 ed603f1
Patch v1.3.1 (#15)
* Changed filename and added MacOS Github Action

* Hardening for MacOS

* Fixing bug in finding OpenVDB libraries

* Increment version
