### 2026-08-18 ifilot/den2obj#14 @ifilot
Thank you for pointing out this bug! I have attempted to patch it and check it via a MacOSX Github Actions workflow. Can you give it another shot and let me know if compilation is now working for you?
