### 2026-08-19 #10100 @WardLT
I'm pleased to recommend this paper be accepted. Excellent documentation and codebase, very clear paper. I'm excited to figure out where I can use it now 😄
