### 2026-08-20 #10100 @jmmshn
^^ Ditto @WardLT
installed and tested the patched version on OSX everything works now!!
