### 2026-08-25 #10100 @ifilot
- I have modified the paper and adjusted the phrasing relating to our earlier work to more accurately reflect the fact that it was based using an older version of the code.
- I have created a tagged release (v1.3.1)
- Created an archive on [Zenodo](https://doi.org/10.5281/zenodo.22095835)
